Only parse the digest challenge part of the response header
It is allowed to provide multiple authentication schemes with optional
challenges in the WWW-Authenticate headers.  This fixes that it always
took the first scheme, and thus failed on a header such as:
  WWW-Authenticate: NTLM, Digest ....

Also frameworks such as Faraday group multiple WWW-Authenticate headers in
a response, for example
  WWW-Authenticate: Digest ...
  WWW-Authenticate: NTLM
into:
  WWW-Authenticate: Digest ..., NTLM
so there is a single way of retrieving header data.

This commit changes the parser to pick out only the digest challenge part
and the authentication parameters until it encounters another scheme (if
any).