Start the nonce count with value 1
The nounce count contains the number of times the client has sent the nonce
value in the request (see Section 3.2.2 of RFC 2617).  This should start
with 1 because the initial request contains the value.

Also, IIS 10.0 seems to ignore requests where nonce count is 0 and resends
a "HTTP 401 Authorization Required" response.